requirements.txt
numpy==1.26.4
matplotlib==3.9.0
scikit-learn==1.5.0
pytorch-lightning==2.2.4
torchmetrics==1.4.0
pyarrow==16.1.0
scipy
pandas
einops==0.7.0
timm==1.0.9

